feat: login and signup routes

+++ src/main/java/com/alemarch/partbin/controllers/ProductController.java
@@ -20,7 +20,7 @@ public class ProductController {
 
 	@GetMapping
 	public Iterable<ProductDto> getProducts(
-			@RequestParam(required = false, name = "categoryId") Long categoryId
+			@RequestParam(required = false, name = "categoryId") Byte categoryId
 	) {
 		var products = (categoryId != null ?
 			productRepository.findByCategory(categoryId) :
